# server/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Aplicație",
    "description": "Description",
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/github/callback": {
      "get": {
        "tags": ["Authentication"],
        "summary": "GitHub Authentication Callback",
        "description": "Callback endpoint for GitHub authentication process.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/login/": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User login",
        "description": "Endpoint to handle user login.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/logout/": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User logout",
        "description": "Endpoint to handle user logout.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/posts/": {
      "post": {
        "tags": ["Posts"],
        "summary": "Create a new post",
        "description": "Endpoint to create a new post.",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "example": "any"
                  },
                  "link": {
                    "example": "any"
                  },
                  "photo": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/users/": {
      "post": {
        "tags": ["Users"],
        "summary": "Create a user profile",
        "description": "Endpoint to create a user profile.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "example": "any"
                  },
                  "lastName": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "github": {
                    "example": "any"
                  },
                  "photo": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Users"],
        "summary": "Get users",
        "description": "Endpoint to get users.",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "array"
                    },
                    "items": {
                      "$ref": "#/components/schemas/User"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "array"
                    },
                    "items": {
                      "$ref": "#/components/schemas/User"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Edit user profile",
        "description": "Endpoint to edit user profile.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "example": "any"
                  },
                  "lastName": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "github": {
                    "example": "any"
                  },
                  "photo": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user profile",
        "description": "Endpoint to delete user profile.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "github": {
            "type": "string"
          },
          "photo": {
            "type": "string"
          }
        }
      },
...
    }
  }
}